move log in/log out to headerComponent
