+++ frontend/src/pages/HomePage.tsx
@@ -6,30 +6,16 @@ import HeaderComponent from "../component/HeaderComponent.tsx";
 import type {Profile} from "../interfaces/Typer.ts";
 
 export default function HomePage() {
-    const [profileId, setProfileId] = useState<string|null>("");
     const [profile, setProfile] = useState<Profile | undefined>(undefined);
 
-    useEffect(() =>{
+    /*useEffect(() =>{
         const stordeUserId = localStorage.getItem("profileId");
         if(!stordeUserId){
             return
         }
-        setProfileId(stordeUserId);
         getProfile(stordeUserId).then(profile => setProfile(profile));
 
-    }, []);
-
-    const navigate = useNavigate();
-    const toLogInPage = () => {
-        navigate("/login");
-    };
-
-    function logUt(){
-        localStorage.setItem("profileId", "");
-        setProfileId("")
-        setProfile(undefined);
-        toLogInPage();
-    }
+    }, []);*/
 
     function consoleProfile(){
         const stordeUserId = localStorage.getItem("profileId");
@@ -43,9 +29,8 @@ export default function HomePage() {
 
     return (
         <>
-            <HeaderComponent logOut={logUt}/>
+            <HeaderComponent />
             <p>Hello: {profile?.profileName}</p>
-            <button onClick={logUt}>Log out</button>
             <button onClick={consoleProfile} >print profile</button>
 
         </>
